backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
pydantic>=2.6.0

# LangChain ecosystem
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.33,<0.4.0
langchain-groq>=0.2.0
langchain-community>=0.3.0,<0.4.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.1.0
langchain-tavily>=0.1.0
langgraph>=0.2.0

# LangSmith
langsmith>=0.2.0

# Vector store & embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0

# HTTP & utils
httpx>=0.27.0
requests>=2.32.0
Pillow>=10.0.0
aiofiles>=24.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
